[#39] Bump Minimum iOS/iPadOS Version to 18.6
