+++ Shared/Views/Assess/Custom Controls/CustomAssessTaps.swift
@@ -19,8 +19,6 @@ extension AssessmentTaker {
             TapButtons.Count()
         case .heartRate:
             TapButtons.HeartRate()
-        default:
-            EmptyView()
         }
     }
